requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
motor==4.3.0                       # Async MongoDB driver
passlib[bcrypt]==1.7.4
python-jose==3.3.0                  # JWT token handling
python-dotenv==1.0.1
llama-index==0.5.6              # Latest llama-index
openai==0.27.8                       # For OpenAI models
anthropic==0.4.1                     # For Claude models
google-generativeai==0.17.0          # For Gemini models
